+++ src/main/java/me/drex/magic_particles/particles/ParticleManager.java
@@ -112,7 +112,6 @@ public class ParticleManager {
                     magicParticle.display(player.createCommandSourceStack());
                 }
             }
-
         }
     }