+++ src/main/java/me/drex/magic_particles/particles/particle/ImageParticle.java
@@ -8,8 +8,8 @@ import me.drex.magic_particles.particles.ParticleManager;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.arguments.EntityAnchorArgument;
 import net.minecraft.core.particles.DustParticleOptions;
+import net.minecraft.network.protocol.game.ClientboundLevelParticlesPacket;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.util.ARGB;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.Display;
@@ -21,6 +21,7 @@ import org.joml.Vector3f;
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
 import java.io.IOException;
+import java.util.function.Consumer;
 
 public class ImageParticle extends AbstractParticle {
 
@@ -43,7 +44,9 @@ public class ImageParticle extends AbstractParticle {
     private final float pixelSize;
     private final Vec3 pos;
     private final Vec2 rotation;
-    private final BufferedImage bufferedImage;
+    private final int width;
+    private final int height;
+    private final int[] colors;
     private final RandomSource random;
 
     protected ImageParticle(String image, float sizeX, float sizeY, float pixelSize, Vec3 pos, EntityAnchorArgument.Anchor anchor, Vec3 origin, Vec2 rotation, Display.BillboardConstraints billboard) {
@@ -55,7 +58,11 @@ public class ImageParticle extends AbstractParticle {
         this.pos = pos;
         this.rotation = rotation;
         try {
-            bufferedImage = ImageIO.read(ParticleManager.PARTICLES_FOLDER.resolve("images").resolve(image).toFile());
+            BufferedImage bufferedImage = ImageIO.read(ParticleManager.PARTICLES_FOLDER.resolve("images").resolve(image).toFile());
+            this.width = bufferedImage.getWidth();
+            this.height = bufferedImage.getHeight();
+            this.colors = new int[this.width * this.height];
+            bufferedImage.getRGB(0, 0, this.width, this.height, this.colors, 0, this.width);
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
@@ -63,12 +70,17 @@ public class ImageParticle extends AbstractParticle {
     }
 
     @Override
-    public void sendParticles(CommandSourceStack source, ServerPlayer player) {
-        float stepX = sizeX / bufferedImage.getWidth();
-        float stepY = sizeY / bufferedImage.getHeight();
-        for (int x = 0; x < bufferedImage.getWidth(); x++) {
-            for (int y = 0; y < bufferedImage.getHeight(); y++) {
-                int color = bufferedImage.getRGB(x, y);
+    public void collectParticlePackets(CommandSourceStack source, Consumer<ClientboundLevelParticlesPacket> collector) {
+        float stepX = sizeX / width;
+        float stepY = sizeY / height;
+
+        Quaternionf billboardRotation = getBillboardRotation(source);
+        Vector3f posVector = pos.toVector3f();
+        Quaternionf rotation = new Quaternionf().rotationYXZ((float) Math.PI - (float) Math.PI / 180 * this.rotation.y, (float) (-Math.PI) / 180 * this.rotation.x, 0.0f);
+
+        for (int x = 0; x < width; x++) {
+            for (int y = 0; y < height; y++) {
+                int color = colors[y * width + x];
 
                 int red = (color & 0xff0000) >> 16;
                 int green = (color & 0xff00) >> 8;
@@ -76,17 +88,17 @@ public class ImageParticle extends AbstractParticle {
                 int alpha = (color & 0xff000000) >>> 24;
 
                 DustParticleOptions particleOptions = new DustParticleOptions(ARGB.color(red, green, blue), pixelSize);
-                float reversedX = bufferedImage.getWidth() - 1 - x;
-                float centeredX = reversedX - ((float) (bufferedImage.getWidth() - 1) / 2);
-                float reversedY = bufferedImage.getHeight() - 1 - y;
-                float centeredY = reversedY - ((float) (bufferedImage.getHeight() - 1) / 2);
+                float reversedX = width - 1 - x;
+                float centeredX = reversedX - ((float) (width - 1) / 2);
+                float reversedY = height - 1 - y;
+                float centeredY = reversedY - ((float) (height - 1) / 2);
 
-                Quaternionf rotation = new Quaternionf().rotationYXZ((float) Math.PI - (float) Math.PI / 180 * this.rotation.y, (float) (-Math.PI) / 180 * this.rotation.x, 0.0f);
-                Vector3f offset = new Vector3f(stepX * centeredX, stepY * centeredY, 0).add(pos.toVector3f());
+                Vector3f offset = new Vector3f(stepX * centeredX, stepY * centeredY, 0).add(posVector);
                 offset = offset.rotate(rotation);
 
-                if (alpha >= random.nextIntBetweenInclusive(1, 255)) {
-                    sendParticles(source, player, particleOptions, false, offset, 1, Vec3.ZERO, 0);
+                // some micro-optimizations to skip random
+                if (alpha >= 255 || (alpha > 0 && alpha >= random.nextIntBetweenInclusive(1, 255))) {
+                    collector.accept(createParticlePacket(source, particleOptions, false, offset, 1, Vec3.ZERO, 0, billboardRotation));
                 }
             }
         }